image-assets: bilingual keywords + English text-search index and tester
Split image asset text into Spanish (SpanishDescription/SpanishKeywords,
feeding the frontend bigram local search) and English (Description/Keywords)
columns. Keywords is a deduplicated joined-word string backing a Sonic
TextSearchColumn for the AI agent's English search; GroupID widened to int32
as the index requires an int32 partition. The sync now reads both
IMAGES_LIST.ES.md and IMAGES_LIST.md and merges them by image ID.

Adds GET.image-asset-text-search and a "Stock Images" tab on /system/testing
that renders the top-10 hits as thumbnails with their relevance weights.

+++ backend/business/types/image_assets_category.go
@@ -4,7 +4,7 @@ import "app/db"
 
 type ImageAssetCategory struct {
 	db.TableStruct[ImageAssetCategoryTable, ImageAssetCategory]
-	GroupID int8   `json:"-"`
+	GroupID int32  `json:"-"`
 	ID      int16  `json:",omitempty"`
 	Name    string `json:",omitempty"`
 	Updated int32  `json:"upd,omitempty"`
@@ -13,7 +13,7 @@ type ImageAssetCategory struct {
 
 type ImageAssetCategoryTable struct {
 	db.TableStruct[ImageAssetCategoryTable, ImageAssetCategory]
-	GroupID db.Col[ImageAssetCategoryTable, int8]
+	GroupID db.Col[ImageAssetCategoryTable, int32]
 	ID      db.Col[ImageAssetCategoryTable, int16]
 	Name    db.Col[ImageAssetCategoryTable, string]
 	Updated db.Col[ImageAssetCategoryTable, int32]
